services/user/create-user/requirements.txt
fastapi
uvicorn
databases[postgresql]
sqlalchemy
passlib[bcrypt]
pydantic[email]
psycopg2-binary
python-dotenv
aiokafka
pytest-asyncio
httpx
asgi-lifespan
aiokafka
loguru
bcrypt==3.2.2
